Move SwitchContainer out of RoomInfoEditView
