+++ app/views/RoomInfoEditView/styles.js
@@ -20,28 +20,6 @@ export default StyleSheet.create({
 	colorDanger: {
 		color: COLOR_DANGER
 	},
-	switchContainer: {
-		flexDirection: 'row',
-		alignItems: 'flex-start'
-	},
-	switchLabelContainer: {
-		flex: 1,
-		paddingHorizontal: 10
-	},
-	switchLabelPrimary: {
-		fontSize: 16,
-		paddingBottom: 6,
-		...sharedStyles.textRegular,
-		...sharedStyles.textColorNormal
-	},
-	switchLabelSecondary: {
-		fontSize: 12,
-		...sharedStyles.textRegular,
-		...sharedStyles.textColorNormal
-	},
-	switch: {
-		alignSelf: 'center'
-	},
 	divider: {
 		height: StyleSheet.hairlineWidth,
 		borderColor: COLOR_SEPARATOR,
