+++ app/views/SwitchContainer/index.js
